Add bottom padding to the site
Previously, the site ended right after the last line of text which
looked weird. Also, for me it is harder to read text at the bottom of
the screen so I like to have some free space on the bottom, so that the
content is more towards the center of the screen.